standarized data set ingestion
This whole notebook should run without error if you have all of the .csv files in the working directory in Colab. 

If it doesn't, please put in an issue for me so I can troubleshoot. This should allow you to ingest any of the csv files and run with no error.